backend/flight-status-backend/requirements.txt
Flask>=2.0.0
jinja2>=3.0.0
kafka-python==2.0.2
pymongo==3.11.4
python-dotenv==0.15.0
six
